Added Custom exceptions for services

+++ ClassBook.BLL/Services/GenericService.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using AutoMapper;
+using ClassBook.BLL.Exceptions;
 using ClassBook.BLL.IServices;
 using ClassBook.DAL;
 using ClassBook.DAL.IRepositories;
@@ -11,7 +12,7 @@ using ClassBook.Domain.Entities;
 
 namespace ClassBook.BLL.Services;
 
-internal class GenericService<T, D, A, U> : IGenericService<T, D, A, U> where T : class where D : class where A : class where U : class
+internal class GenericService<T, D, U> : IGenericService<T, D, U> where T : class where D : class where U : class
 {
     internal readonly IGenericRepository<T> Repository;
     internal readonly IMapper Mapper;
@@ -25,24 +26,27 @@ internal class GenericService<T, D, A, U> : IGenericService<T, D, A, U> where T 
     public async Task<IEnumerable<D>> GetAllAsync()
     {
         var list = await Repository.GetAllAsync();
+
         return Mapper.Map<IEnumerable<D>>(list);
     }
 
     public async Task<D> GetByIdAsync(int id)
     {
         var o = await Repository.GetByIdAsync(id);
+
+        if (o == null)
+            throw new NotFoundException();
+
         return Mapper.Map<D>(o);
     }
 
-    public async Task AddAsync(A obj)
+    public async Task UpdateAsync(int id, U obj)
     {
-        var help = Mapper.Map<T>(obj);
-        await Repository.InsertAsync(help);
-        await Repository.SaveAsync();
-    }
+        var o = await Repository.GetByIdAsync(id);
 
-    public async Task UpdateAsync(U obj)
-    {
+        if (o == null)
+            throw new NotFoundException();
+
         var help = Mapper.Map<T>(obj);
         await Repository.UpdateAsync(help);
         await Repository.SaveAsync();
@@ -50,6 +54,11 @@ internal class GenericService<T, D, A, U> : IGenericService<T, D, A, U> where T 
 
     public async Task RemoveAsync(int id)
     {
+        var o = await Repository.GetByIdAsync(id);
+
+        if (o == null)
+            throw new NotFoundException();
+
         await Repository.DeleteAsync(id);
         await Repository.SaveAsync();
     }
